feat(backEnd): unification of dashboard endpoints
