--- backend/app/Http/Controllers/Api/Dashboard/StudentsController.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Api\Dashboard;
-
-use App\Http\Controllers\Controller;
-use App\Models\User;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Http\Request;
-
-class StudentsController extends Controller
-{
-
-    public function studentCountPerSubArea(Request $request): array
-    {
-        $data = $request->validate([
-            'selectedFilter' => 'nullable|string|in:mestrando,doutorando,completed',
-        ]);
-
-        $filter = $data['selectedFilter'] ?? null;
-
-        return User::userCountPerSubArea($filter);
-    }
-
-    public function studentCountPerArea(Request $request): array
-    {
-        //return ['fields' => ['CG', 'Análise de Dados', 'I.A',],
-        //        'data' => [12, 3, 5,]];
-
-        $data = $request->validate([
-            'selectedFilter' => 'nullable|string|in:mestrando,doutorando,completed',
-        ]);
-
-        $filter = $data['selectedFilter'] ?? null;
-
-        return User::userCountPerArea($filter);
-    }
-
-}
